+++ Search/Adapter/ZendLuceneAdapter.php
@@ -1,13 +1,23 @@
 <?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
 
 namespace Massive\Bundle\SearchBundle\Search\Adapter;
 
-use ZendSearch\Lucene;
 use Massive\Bundle\SearchBundle\Search\AdapterInterface;
 use Massive\Bundle\SearchBundle\Search\Document;
+use Massive\Bundle\SearchBundle\Search\Factory;
 use Massive\Bundle\SearchBundle\Search\Field;
 use Massive\Bundle\SearchBundle\Search\QueryHit;
 use Symfony\Component\Finder\Finder;
+use ZendSearch\Lucene;
+use Massive\Bundle\SearchBundle\Search\SearchQuery;
 
 /**
  * Adapter for the ZendSearch library
@@ -16,34 +26,40 @@ use Symfony\Component\Finder\Finder;
  * http://framework.zend.com/manual/1.12/en/zend.search.lucene.html 
  *   (docs for 1.2 version apply equally to 2.0)
  *
+ * Note this adapter implements localization by creating an index for each
+ * locale if a locale is specified.
+ *
  * @author Daniel Leech <daniel@massive.com>
  */
 class ZendLuceneAdapter implements AdapterInterface
 {
     const ID_FIELDNAME = '__id';
+    const CLASS_TAG = '__class';
+
+    // TODO: This fields should be handled at a higher level
     const URL_FIELDNAME = '__url';
     const TITLE_FIELDNAME = '__title';
     const DESCRIPTION_FIELDNAME = '__description';
-    const CLASS_TAG = '__class';
+    const IMAGE_URL = '__image_url';
 
+    /**
+     * The base directory for the search indexes
+     * @var string
+     */
     protected $basePath;
+
+    /**
+     * @var \Massive\Bundle\SearchBundle\Search\Factory
+     */
+    protected $factory;
 
     /**
      * @param string $basePath Base filesystem path for the index
      */
-    public function __construct($basePath)
+    public function __construct(Factory $factory, $basePath)
     {
         $this->basePath = $basePath;
-    }
-
-    /**
-     * Determine the index path for a given index name
-     * @param string $indexName
-     * @return string
-     */
-    protected function getIndexPath($indexName)
-    {
-        return sprintf('%s/%s', $this->basePath, $indexName);
+        $this->factory = $factory;
     }
 
     /**
@@ -51,26 +67,26 @@ class ZendLuceneAdapter implements AdapterInterface
      */
     public function index(Document $document, $indexName)
     {
-        $indexPath = $this->getIndexPath($indexName);
+        $index = $this->getLuceneIndex($document, $indexName);
 
-        if (!file_exists($indexPath)) {
-            $index = Lucene\Lucene::create($indexPath);
-        } else {
-            $index = Lucene\Lucene::open($indexPath);
-
-            // check to see if the subject already exists
-            $this->removeExisting($index, $document);
-        }
+        // check to see if the subject already exists
+        $this->removeExisting($index, $document);
 
         $luceneDocument = new Lucene\Document();
 
         foreach ($document->getFields() as $field) {
             switch ($field->getType()) {
                 case Field::TYPE_STRING:
-                default:
-                    $luceneDocument->addField(Lucene\Document\Field::Text($field->getName(), $field->getValue()));
+                    $luceneField = Lucene\Document\Field::Text($field->getName(), $field->getValue());
                     break;
+                default:
+                    throw new \InvalidArgumentException(sprintf(
+                        'Search field type "%s" is not know. Known types are: %s',
+                        implode(', ', Field::getValidTypes())
+                    ));
             }
+
+            $luceneDocument->addField($luceneField);
         }
 
         // add meta fields - used internally for showing the search results, etc.
@@ -79,34 +95,41 @@ class ZendLuceneAdapter implements AdapterInterface
         $luceneDocument->addField(Lucene\Document\Field::Keyword(self::TITLE_FIELDNAME, $document->getTitle()));
         $luceneDocument->addField(Lucene\Document\Field::Keyword(self::DESCRIPTION_FIELDNAME, $document->getDescription()));
         $luceneDocument->addField(Lucene\Document\Field::Keyword(self::CLASS_TAG, $document->getClass()));
+        $luceneDocument->addField(Lucene\Document\Field::Keyword(self::IMAGE_URL, $document->getImageUrl()));
 
         $index->addDocument($luceneDocument);
     }
 
     /**
-     * Remove the existing entry for the given Document from the index, if it exists.
-     *
-     * @param Lucene\Index $index The Zend Lucene Index
-     * @param Document $document The Massive Search Document
+     * {@inheritDoc}
      */
-    protected function removeExisting(Lucene\Index $index, Document $document)
+    public function deindex(Document $document, $indexName)
     {
-        $hits = $index->find(self::ID_FIELDNAME . ':' . $document->getId());
-
-        foreach ($hits as $hit) {
-            $index->delete($hit->id);
-        }
+        $index = $this->getLuceneIndex($document, $indexName);
+        $this->removeExisting($index, $document);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function search($queryString, array $indexNames = array())
+    public function search(SearchQuery $searchQuery)
     {
+        $indexNames = $searchQuery->getIndexes();
+        $locale = $searchQuery->getLocale();
+        $queryString = $searchQuery->getQueryString();
+
         $searcher = new Lucene\MultiSearcher();
 
         foreach ($indexNames as $indexName) {
-            $searcher->addIndex(Lucene\Lucene::open($this->getIndexPath($indexName)));
+            $indexPath = $this->getIndexPath(
+                $this->getLocalizedIndexName($indexName, $locale)
+            );
+
+            if (!file_exists($indexPath)) {
+                continue;
+            }
+
+            $searcher->addIndex(Lucene\Lucene::open($indexPath));
         }
 
         $query = Lucene\Search\QueryParser::parse($queryString);
@@ -116,8 +139,9 @@ class ZendLuceneAdapter implements AdapterInterface
         $hits = array();
 
         foreach ($luceneHits as $luceneHit) {
-            $hit = new QueryHit();
-            $document = new Document();
+            $hit = $this->factory->makeQueryHit();
+            $document = $this->factory->makeDocument();
+
             $hit->setDocument($document);
             $hit->setScore($luceneHit->score);
 
@@ -129,11 +153,12 @@ class ZendLuceneAdapter implements AdapterInterface
             $document->setDescription($luceneDocument->getFieldValue(self::DESCRIPTION_FIELDNAME));
             $document->setUrl($luceneDocument->getFieldValue(self::URL_FIELDNAME));
             $document->setClass($luceneDocument->getFieldValue(self::CLASS_TAG));
+            $document->setImageUrl($luceneDocument->getFieldValue(self::IMAGE_URL));
 
             $hit->setId($document->getId());
 
             foreach ($luceneDocument->getFieldNames() as $fieldName) {
-                $document->addField(Field::create($fieldName, $luceneDocument->getFieldValue($fieldName)));
+                $document->addField($this->factory->makeField($fieldName, $luceneDocument->getFieldValue($fieldName)));
             }
             $hits[] = $hit;
         }
@@ -169,5 +194,64 @@ class ZendLuceneAdapter implements AdapterInterface
         }
 
         return $status;
+    }
+
+    /**
+     * Returns the lucene index for the locale determined by the document
+     * @param Document $document
+     * @param $indexName
+     * @return Lucene\SearchIndexInterface
+     */
+    private function getLuceneIndex(Document $document, $indexName)
+    {
+        $locale = $document->getLocale();
+        $indexName = $this->getLocalizedIndexName($indexName, $locale);
+        $indexPath = $this->getIndexPath($indexName);
+
+        if (!file_exists($indexPath)) {
+             Lucene\Lucene::create($indexPath);
+        }
+
+        return Lucene\Lucene::open($indexPath);
+    }
+
+    /**
+     * Determine the index path for a given index name
+     * @param string $indexName
+     * @return string
+     */
+    private function getIndexPath($indexName)
+    {
+        return sprintf('%s/%s', $this->basePath, $indexName);
+    }
+
+    /**
+     * Return the localized index name if the locale
+     * is given.
+     *
+     * @return string
+     */
+    private function getLocalizedIndexName($indexName, $locale = null)
+    {
+        if (null === $locale) {
+            return $indexName;
+        }
+
+        return $indexName . '_' . $locale;
+    }
+
+    /**
+     * Remove the existing entry for the given Document from the index, if it exists.
+     *
+     * @param Lucene\Index $index The Zend Lucene Index
+     * @param Document $document The Massive Search Document
+     */
+    private function removeExisting(Lucene\Index $index, Document $document)
+    {
+        $hits = $index->find(self::ID_FIELDNAME . ':' . $document->getId());
+
+        foreach ($hits as $hit) {
+            $index->delete($hit->id);
+        }
     }
 }