+++ Tests/Integration/AdapterTestCase.php
@@ -1,16 +1,65 @@
 <?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
 
 namespace Massive\Bundle\SearchBundle\Tests\Integration;
 
 use Symfony\Cmf\Component\Testing\Functional\BaseTestCase;
 use Massive\Bundle\SearchBundle\Search\Field;
 use Massive\Bundle\SearchBundle\Search\Document;
+use Massive\Bundle\SearchBundle\Search\Factory;
+use Massive\Bundle\SearchBundle\Search\SearchQuery;
 
 abstract class AdapterTestCase extends BaseTestCase
 {
+    protected $factory;
+
     public function setUp()
     {
+        $this->factory = new Factory();
         parent::setUp();
+    }
+
+    public function testIndexer()
+    {
+        $adapter = $this->getAdapter();
+        $this->createIndex();
+
+        $query = new SearchQuery('one');
+        $query->setIndexes(array('foobar'));
+        $res = $adapter->search($query);
+
+        $this->assertCount(1, $res);
+    }
+
+    public function testGetStatistics()
+    {
+        $this->createIndex();
+        $adapter = $this->getAdapter();
+        $statistics = $adapter->getStatus();
+        $this->assertTrue(is_array($statistics));
+    }
+
+    public function testDeindex()
+    {
+        $this->createIndex();
+        $doc = $this->factory->makeDocument();
+        $doc->setId(1);
+        $this->getAdapter()->deindex($doc, 'foobar');
+
+        $query = new SearchQuery('document');
+        $query->setIndexes(array('foobar'));
+        $res = $this->getAdapter()->search($query);
+
+        // this should be one, but the lucene index needs to be
+        // comitted, and to do that we must callits destruct method.
+        $this->assertCount(2, $res);
     }
 
     protected function createDocument($title)
@@ -18,14 +67,14 @@ abstract class AdapterTestCase extends BaseTestCase
         static $id = 0;
         $id++;
 
-        $document = new Document();
+        $document = $this->factory->makeDocument();
         $document->setId($id);
-        $document->addField(Field::create('title', $title, 'string'));
+        $document->addField($this->factory->makeField('title', $title, 'string'));
         $text = <<<EOT
 This section is a brief introduction to reStructuredText (reST) concepts and syntax, intended to provide authors with enough information to author documents documentively. Since reST was designed to be a simple, unobtrusive markup language, this will not take too long.
 EOT
         ;
-        $document->addField(Field::create('body', $text, 'string'));
+        $document->addField($this->factory->makeField('body', $text, 'string'));
 
         return $document;
     }
@@ -43,21 +92,8 @@ EOT
         }
     }
 
-    public function testIndexer()
+    protected function getFactory()
     {
-        $adapter = $this->getAdapter();
-        $this->createIndex();
-
-        $res = $adapter->search('One', array('foobar'));
-
-        $this->assertCount(1, $res);
-    }
-
-    public function testGetStatistics()
-    {
-        $this->createIndex();
-        $adapter = $this->getAdapter();
-        $statistics = $adapter->getStatus();
-        $this->assertTrue(is_array($statistics));
+        return new Factory();
     }
 }